feat: make settings key add generic
